requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
alembic==1.13.2
python-dotenv==1.0.1
waitress==3.0.0
pyodbc==5.1.0
greenlet>=3.0
fpdf2>=2.7.8    # pdf生成器
